updating create streamer
