+++ x/streamer/keeper/iterator.go
@@ -1,7 +1,10 @@
 package keeper
 
 import (
+	"encoding/json"
+
 	"github.com/dymensionxyz/dymension/x/streamer/types"
+	db "github.com/tendermint/tm-db"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
 )
@@ -30,4 +33,25 @@ func (k Keeper) ActiveStreamsIterator(ctx sdk.Context) sdk.Iterator {
 // FinishedStreamsIterator returns the iterator for all finished streams.
 func (k Keeper) FinishedStreamsIterator(ctx sdk.Context) sdk.Iterator {
 	return k.iterator(ctx, types.KeyPrefixFinishedStreams)
+}
+
+// getStreamsFromIterator iterates over everything in a stream's iterator, until it reaches the end. Return all streams iterated over.
+func (k Keeper) getStreamsFromIterator(ctx sdk.Context, iterator db.Iterator) []types.Stream {
+	streams := []types.Stream{}
+	defer iterator.Close() // nolint: errcheck
+	for ; iterator.Valid(); iterator.Next() {
+		streamIDs := []uint64{}
+		err := json.Unmarshal(iterator.Value(), &streamIDs)
+		if err != nil {
+			panic(err)
+		}
+		for _, streamID := range streamIDs {
+			stream, err := k.GetStreamByID(ctx, streamID)
+			if err != nil {
+				panic(err)
+			}
+			streams = append(streams, *stream)
+		}
+	}
+	return streams
 }
